EmployeeApp/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
